moves visualize_meshes to base.py
