+++ menpowidgets/menpo3dwidgets/__init__.py
@@ -1,2 +1,1 @@
 from .base import visualize_shape_model_3d, visualize_morphable_model
-from .meshes import visualize_meshes
